refactor: move classes to config-loader
