--- packages/classes/index.js
@@ -1,7 +1,0 @@
-const ReadFile = require('./src/read-file');
-const EikConfig = require('./src/eik-config');
-
-module.exports = {
-    ReadFile,
-    EikConfig,
-};
