--- packages/classes/test/eik.json
@@ -1,10 +1,0 @@
-{
-  "name": "my-app",
-  "version": "1.0.0",
-  "server": "https://assets.myeikserver.com",
-  "files": {
-    "esm.js": "./assets/esm.js",
-    "esm.css": "./assets/esm.css",
-    "/assets": "./assets/**/*.map"
-  }
-}
