+++ packages/config-loader/src/classes/file-mapping.js
