+++ packages/config-loader/src/classes/resolved-files.js
