+++ packages/config-loader/src/errors/custom-error.js
