+++ packages/config-loader/test/local-file-location/contentType.test.js
@@ -1,5 +1,5 @@
 const tap = require('tap');
-const LocalFileLocation = require('../../src/internal/local-file-location.js');
+const LocalFileLocation = require('../../src/classes/local-file-location.js');
 
 tap.test('LocalFileLocation: .contentType for ./my/file.json', (t) => {
     const subject = new LocalFileLocation('./my/file.json', __dirname);
