+++ packages/schemas/package.json
@@ -2,7 +2,7 @@
   "name": "@eik/common-schemas",
   "version": "1.0.0",
   "description": "Shared schema logic for Eik packages",
-  "main": "src/index.js",
+  "main": "index.js",
   "scripts": {
     "style:check": "prettier -c .",
     "style:format": "prettier -w .",
@@ -19,7 +19,8 @@
     "ajv": "8.11.0",
     "ajv-formats": "2.1.1",
     "semver": "7.3.7",
-    "validate-npm-package-name": "4.0.0"
+    "validate-npm-package-name": "4.0.0",
+    "tap": "15.2.3"
   },
   "devDependencies": {
     "@types/node": "18.0.6",
@@ -30,7 +31,6 @@
     "eslint-plugin-prettier": "4.2.1",
     "@babel/eslint-parser": "7.16.5",
     "prettier": "2.5.1",
-    "tap": "15.2.3",
     "typescript": "4.7.4",
     "json-schema-to-typescript": "10.1.5"
   }
